refactor(aux): extract R2 error-predicates leaf into auxiliary_client_errors.py
Byte-verbatim extraction of agent/auxiliary_client.py lines 3615-4017
(R2 slice, epic #78647 / #78635) into agent/auxiliary_client_errors.py:
the error-classification predicates, the recently-402'd unhealthy-provider
cache, and the transient-retry constants — a self-contained leaf family
with zero intra-file function deps.

Seam identity: agent/auxiliary_client.py re-exports all 22 moved names,
so every existing import / lazy-import / monkeypatch target resolves to
the same objects (getattr identity verified 22/22 in
tests/agent/test_aux_error_predicates_seam.py).

Interlock note: no open PR touches any byte inside the window (adjacent
edits only: #79792 above, #79826 ~370 lines above); the window is
byte-intact at the pin (sha256 4bf08251...), so this extraction is safe
to land before or after those PRs. Sequence after their merge or
interlock in the manifest.
